DB_tools/requirements.txt
openai
mysql-connector-python
bibtexparser
